requirements.txt
asgiref==3.4.1
chardet==4.0.0
dj-database-url==0.5.0
Django==4.0
django-heroku==0.3.1
djangorestframework==3.13.1
Pillow==9.0.0
psycopg2==2.9.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
